Fixed a serious bug in libabf.
git-svn-id: svn://mosedal.net/abf/code/trunk/libabf@222 3c19b6e9-aa5f-dd11-88a4-001302d38fef

+++ libabf.cpp
@@ -109,7 +109,7 @@ fseek(fout, HeaderSize, SEEK_SET);
 int Size = 0;
 for (unsigned short i = 0; i < _NumSections; i++) {
 fwrite(&i, sizeof(short), 1, fout);
-fwrite(&Size, sizeof(short), 1, fout);
+fwrite(&Size, sizeof(int), 1, fout);
 }
 }
 void AbfEncoder::WriteSection() {
